refactor: agents
