+++ agents/tools.go
@@ -1,0 +1,10 @@
+package agents
+
+import (
+	"context"
+)
+
+type Tool interface {
+	Name() string
+	Call(ctx context.Context, input string) (string, error)
+}
